backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0
PyMySQL==1.1.0
python-dotenv==1.0.0
Flask-Cors==4.0.0
